+++ Makefile
@@ -2,7 +2,7 @@
 TARGET = kk_project
 
 # 소스 파일 목록
-SRC = main.cpp upper_body.cpp
+SRC = main.cpp body.cpp head.cpp leg.cpp
 
 # 컴파일러 및 옵션
 CXX = g++
