+++ head.h
@@ -1,0 +1,6 @@
+#ifndef HEAD_H
+#define HEAD_H
+
+void drawKKHead(); // 머리 전체 그리기 함수 선언
+
+#endif
