+++ leg.h
@@ -1,0 +1,7 @@
+// leg.h
+#ifndef LEG_H
+#define LEG_H
+
+void drawFootWithToes(float xOffset); // 다리(발) 그리기 함수 선언
+
+#endif
